backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
httpx==0.27.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
langgraph>=0.2.34
sqlalchemy==2.0.23
